Logout + auth tokens in cookies + autorefresh
Autorefresh is disabled for now. The process is working (new token and refresh token are generated and stored in cookies when a refresh token is detected in cookie jar without a token), but the original request can't be restarted with proper auth server-side.
Depending of the final choice for auth, it could be easier to do it in a cleaner way from frontend (if a 401 is thrown on first request, a /token/refresh request is issued, and if it's successful the original request is sent again)